+++ src/app/components/Feedback/Feedback.jsx
@@ -2,7 +2,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import FocusTrap from 'focus-trap-react';
 import axios from 'axios';
-import { Button, ButtonTypes, Input, Label, HelperErrorText, Link } from '@nypl/design-system-react-components';
+import {
+  Button,
+  HelperErrorText,
+  Label,
+  Link,
+  TextInput
+} from '@nypl/design-system-react-components';
 
 import { trackDiscovery } from '../../utils/utils';
 import appConfig from '../../data/appConfig';
@@ -93,16 +99,15 @@ class Feedback extends React.Component {
     const { submit } = this.props;
 
     return (
-      <div className="feedback nypl-ds">
+      <div className="feedback">
         <Button
+          aria-haspopup='true'
+          aria-expanded={showForm}
+          aria-controls='feedback-menu'
+          buttonType="secondary"
+          id="help-feedback"
           className="feedback-button"
           onClick={() => this.toggleForm()}
-          attributes={{
-            'aria-haspopup': 'true',
-            'aria-expanded': showForm,
-            'aria-controls': 'feedback-menu',
-          }}
-          buttonType={ButtonTypes.Secondary}
         >
           Help & Feedback
         </Button>
@@ -126,7 +131,7 @@ class Feedback extends React.Component {
                   onSubmit={this.onSubmitForm}
                 >
                   <div>
-                    <Label htmlFor="feedback-textarea-comment">
+                    <Label htmlFor="feedback-textarea-comment" id="feedback-textarea">
                       Comments*
                     </Label>
                     <textarea
@@ -141,52 +146,51 @@ class Feedback extends React.Component {
                     />
                     <HelperErrorText
                       id="helper-text"
-                      isError={commentInputError}
-                    >
-                      {commentInputError ? 'Please fill out this field' : ''}
-                    </HelperErrorText>
+                      isInvalid={commentInputError}
+                      text={commentInputError ? 'Please fill out this field' : ''}
+                    />
                   </div>
                   <div>
-                    <Label htmlFor="feedback-input-email">
-                      Email <span>(required if you would like a response from us)</span>
-                    </Label>
-                    <Input
-                      required
-                      attributes={{
-                        name: 'email',
-                        onChange: this.handleInputChange,
-                        ref: this.emailInput,
-                      }}
+                    <TextInput
                       id="feedback-input-email"
+                      labelText={
+                        <>Email <span>(required if you would like a response from us)</span></>
+                      }
+                      name='email'
+                      onChange={this.handleInputChange}
+                      ref={this.emailInput}
+                      // isRequired
                       type="email"
                       value={fields.email}
+                      marginTop="l"
                     />
                   </div>
                   <div className="privacy-policy">
                     <Link
                       href="https://www.nypl.org/help/about-nypl/legal-notices/privacy-policy"
                       target="_blank"
-                    >Privacy Policy
+                    >
+                      Privacy Policy
                     </Link>
                   </div>
                   <Button
-                    type="reset"
+                    aria-expanded={!showForm}
+                    aria-controls='feedback-menu'
+                    buttonType="secondary"
+                    id="cancel-feedback"
                     className={`cancel-button ${!showForm ? 'hidden' : ''}`}
                     onClick={e => this.deactivateForm(e)}
-                    attributes={{
-                      'aria-expanded': !showForm,
-                      'aria-controls': 'feedback-menu',
-                    }}
-                    buttonType={ButtonTypes.Secondary}
+                    type="reset"
                   >
                     Cancel
                   </Button>
 
                   <Button
-                    type="submit"
-                    buttonType={ButtonTypes.Primary}
                     className="submit-button"
-                  >Submit
+                    id="submit-feedback"
+                    type="submit"
+                  >
+                    Submit
                   </Button>
                 </form>
               </>