+++ src/app/components/Redirect404/Redirect404.jsx
@@ -21,9 +21,9 @@ const Redirect404 = (props, context) => {
 
   return (
     <DocumentTitle title={`404 | ${appConfig.displayTitle} | NYPL`}>
-      <div className="redirect404 .nypl-ds">
+      <div className="redirect404">
         <div className="redirect404text">
-          <Heading level={1}>
+          <Heading level="one">
             We&apos;re Sorry...
           </Heading>
           <p>