+++ src/app/pages/BibPage.jsx
@@ -22,6 +22,8 @@ import {
   compressSubjectLiteral,
   getGroupedNotes,
   getIdentifiers,
+  stringDirection,
+  matchParallels,
 } from '../utils/bibDetailsUtils';
 import appConfig from '../data/appConfig';
 import { itemBatchSize } from '../data/constants';
@@ -74,9 +76,10 @@ const checkForMoreItems = (bib, dispatch) => {
 };
 
 export const BibPage = (
-  { bib, location, searchKeywords, dispatch, resultSelection },
+  { bib, location, searchKeywords, dispatch, resultSelection, features },
   context,
 ) => {
+  const useParallels = features && features.includes('parallels')
   if (!bib || parseInt(bib.status, 10) === 404) {
     return <BibNotFound404 context={context} />;
   }
@@ -106,12 +109,13 @@ export const BibPage = (
   // Make a copy of the `bib` so we can add additional fields with
   // computed data values that will make rendering them easier in
   // the `BibDetails` component.
-  const newBibModel = { ...bib };
-  newBibModel['notesGroupedByNoteType'] = getGroupedNotes(bib);
+  const newBibModel = matchParallels(bib, useParallels);
+  newBibModel['notesGroupedByNoteType'] = getGroupedNotes(newBibModel);
   newBibModel['owner'] = getOwner(bib);
   newBibModel['updatedIdentifiers'] = getIdentifiers(newBibModel, bottomFields);
   newBibModel['updatedSubjectLiteral'] = compressSubjectLiteral(bib);
 
+  const mainHeading = [ bib.parallelTitle, bib.title, [' ']].reduce((acc, el) => acc || (el && el.length && el[0]), null);
   return (
     <RouterProvider value={context}>
       <SccContainer
@@ -119,11 +123,9 @@ export const BibPage = (
         className='nypl-item-details'
         pageTitle='Item Details'
       >
-        <section className='nypl-item-details__heading'>
-          <Heading level={2}>
-            {newBibModel.title && newBibModel.title.length
-              ? newBibModel.title[0]
-              : ' '}
+        <section className='nypl-item-details__heading' dir={stringDirection(mainHeading, useParallels)}>
+          <Heading level="two">
+            { mainHeading }
           </Heading>
           <BackToSearchResults result={resultSelection} bibId={bibId} />
         </section>
@@ -136,22 +138,27 @@ export const BibPage = (
               items,
             )}
             fields={topFields}
+            features={features}
           />
         </section>
 
-        {items.length && !isElectronicResources && (
-          <section style={{ marginTop: '20px' }}>
-            <ItemsContainer
-              key={bibId}
-              shortenItems={location.pathname.indexOf('all') !== -1}
-              items={items}
-              bibId={bibId}
-              itemPage={location.search}
-              searchKeywords={searchKeywords}
-              holdings={newBibModel.holdings}
-            />
-          </section>
-        )}
+        {
+          items.length && !isElectronicResources ?
+          (
+            <section style={{ marginTop: '20px' }}>
+              <ItemsContainer
+                key={bibId}
+                shortenItems={location.pathname.indexOf('all') !== -1}
+                items={items}
+                bibId={bibId}
+                itemPage={location.search}
+                searchKeywords={searchKeywords}
+                holdings={newBibModel.holdings}
+              />
+            </section>
+          ) :
+          null
+        }
 
         {newBibModel.holdings && (
           <section style={{ marginTop: '20px' }}>
@@ -160,7 +167,7 @@ export const BibPage = (
         )}
 
         <section style={{ marginTop: '20px' }}>
-          <Heading level={3}>Details</Heading>
+          <Heading level="three">Details</Heading>
           <BibDetails
             additionalData={
               isNyplBnumber(newBibModel.uri) && newBibModel.annotatedMarc
@@ -170,6 +177,7 @@ export const BibPage = (
             bib={newBibModel}
             electronicResources={aggregatedElectronicResources}
             fields={bottomFields}
+            features={features}
           />
         </section>
 
@@ -188,6 +196,7 @@ BibPage.propTypes = {
   bib: PropTypes.object,
   dispatch: PropTypes.func,
   resultSelection: PropTypes.object,
+  features: PropTypes.array,
 };
 
 BibPage.contextTypes = {
@@ -202,6 +211,7 @@ const mapStateToProps = ({
   page,
   sortBy,
   resultSelection,
+  features,
 }) => ({
   bib,
   searchKeywords,
@@ -210,6 +220,7 @@ const mapStateToProps = ({
   page,
   sortBy,
   resultSelection,
+  features,
 });
 
 export default withRouter(connect(mapStateToProps)(BibPage));